Created misc/holography package
